restart megacmd after update in linux

+++ build/templates/megacmd/megacmd.spec
@@ -346,9 +346,21 @@ sysctl -p /etc/sysctl.d/100-megacmd-inotify-limit.conf
 ### END of POSTINST
 
 
+%preun
+[ "$1" == "1" ] && killall -s SIGUSR1 mega-cmd-server 2> /dev/null || true
+
+
 %postun
-killall mega-cmd 2> /dev/null || true
-killall mega-cmd-server 2> /dev/null || true
+
+# kill running MEGAcmd instance when uninstall (!upgrade)
+[ "$1" == "0" ] && killall mega-cmd 2> /dev/null || true
+[ "$1" == "0" ] && killall mega-cmd-server 2> /dev/null || true
+
+
+%posttrans
+# to restore dormant MEGAcmd upon updates
+killall -s SIGUSR2 mega-cmd-server 2> /dev/null || true
+
 
 %clean
 %{?buildroot:%__rm -rf "%{buildroot}"}
